Shell Recorder: Add option to keep protocol file

+++ tests/shell/shell_recorder/shell_recorder.sh
@@ -230,6 +230,18 @@ run_script()
 
 trap cleanup EXIT INT QUIT TERM
 
+# Parse optional argument `-p`
+OPTIND=1
+keepProtocol='false'
+while getopts "p" opt; do
+	case "$opt" in
+	p)
+		keepProtocol='true'
+		;;
+	esac
+done
+shift $((OPTIND-1))
+
 FILE=$1
 Mountpoint=
 DBFile=
@@ -254,7 +266,8 @@ nbTest=0
 
 if [ "$#" -lt '1' ] || [ "$#" -gt '2' ];
 then
-	printf 'Usage: %s input_script [protocol to compare]\n' "$0"
+	printf 'Usage: %s [-p] input_script [protocol to compare]\n\n' "$0"
+	printf '       -p    keep protocol file\n' "$0"
 	rm "$OutFile"
 	exit 0
 fi
@@ -286,7 +299,7 @@ then
 	fi
 fi
 
-if [ "$EVAL" -eq 0 ]; then
+if [ "$EVAL" -eq 0 ] && [ $keepProtocol == 'false' ]; then
 	rm -f "$OutFile"
 else
 	>&2 printf '\n📕  Protocol File: %s\n' "$OutFile"
